requirements.txt
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
idna==2.10
iniconfig==1.1.1
marshmallow==3.8.0
packaging==20.4
peewee==3.14.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.3
pytest==6.1.2
pytest-cov==2.10.1
pytest-mock==3.3.1
python-dotenv==0.15.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.25.11